refactorisation et réorganisation
