+++ public/manifest.json
@@ -14,13 +14,13 @@
   "permissions": ["activeTab", "scripting", "downloads", "notifications"],
   "host_permissions": ["<all_urls>"],
   "background": {
-    "service_worker": "background.js",
+    "service_worker": "src/background/index.js",
     "type": "module"
   },
   "content_scripts": [
     {
       "matches": ["<all_urls>"],
-      "js": ["contentScript.js"],
+      "js": ["src/content/index.js"],
       "run_at": "document_idle"
     }
   ],
