--- public/test-data.json
@@ -1,9 +1,0 @@
-[
-  {"url": "https://images-dynamic-arcteryx.imgix.net/details/1350x1710/S25-X000009556-Atom-Hoody-Void-Front-View.jpg?auto=format%2Ccompress&q=75&fill=white", "format": "avif", "weight": 171.99},
-  {"url": "https://images-dynamic-arcteryx.imgix.net/details/1350x1710/S25-X000009556-Atom-Hoody-Void-Back-View.jpg?auto=format%2Ccompress&q=75&fill=white", "format": "avif", "weight": 225.8},
-  {"url": "https://images-dynamic-arcteryx.imgix.net/details/1350x1710/S25-X000009556-Atom-Hoody-Void-Side-View.jpg?auto=format%2Ccompress&q=75&fill=white", "format": "avif", "weight": 137.55},
-  {"url": "https://images-dynamic-arcteryx.imgix.net/details/1350x1710/S25-X000009556-Atom-Hoody-Void-Full-Body.jpg?auto=format%2Ccompress&q=75&fill=white", "format": "avif", "weight": 160.75},
-  {"url": "https://images-dynamic-arcteryx.imgix.net/details/1350x1710/S25-X000009556-Atom-Hoody-Void-Detail-1.jpg?auto=format%2Ccompress&q=75&fill=white", "format": "avif", "weight": 235.99},
-  {"url": "https://images-dynamic-arcteryx.imgix.net/details/1350x1710/S25-X000009556-Atom-Hoody-Void-Detail-2.jpg?auto=format%2Ccompress&q=75&fill=white", "format": "avif", "weight": 309.31},
-  {"url": "https://images-dynamic-arcteryx.imgix.net/details/1350x1710/S25-X000009556-Atom-Hoody-Void-Front-View.jpg?auto=format%2Ccompress&q=70&fit=crop&fill=white&ixlib=react-9.8.1&w=927", "format": "avif", "weight": 67.64}
-]
