+++ src/background/panelManager.js
@@ -1,0 +1,102 @@
+// Panel management for Chrome extension
+// Handles UI panel injection and toggle functionality
+
+// Fonction injectée dans la page pour toggler le panneau
+export function toggleTTO() {
+  const containerId = 'tto-extension-container';
+  const ANIMATION_DELAY = 50; // délai avant slide-in (ms)
+
+  // Ouvre le panneau
+  function openPanel() {
+
+    const oldScript = document.querySelector('script[src*="main.js"]');
+    if (oldScript) {
+      oldScript.remove();
+    }
+
+    const oldStyle = document.querySelector('link[href*="index-"]');
+    if (oldStyle) {
+      oldStyle.remove();
+    }
+
+    const styleLink = document.createElement('link');
+    styleLink.rel = 'stylesheet';
+    // Charger le CSS sans hash généré index.css
+    styleLink.href = chrome.runtime.getURL('assets/index.css');
+    document.head.appendChild(styleLink);
+
+    const container = document.createElement('div');
+    container.id = containerId;
+    Object.assign(container.style, {
+      position: 'fixed',
+      top: '0',
+      right: '0',
+      width: '515px',
+      maxHeight: '70%',
+      zIndex: '999999',
+      backgroundColor: '#fff',
+      borderRadius: '0px 0px 0px 15px',
+      boxShadow: '-5px 5px 10px rgba(0, 0, 0, 0.15)',
+      overflow: 'auto',
+      transform: 'translateX(100%)',
+      transition: 'transform 0.6s cubic-bezier(.5,1.6,.4,1)'
+    });
+
+    const appRoot = document.createElement('div');
+    appRoot.id = 'tto-extension-root';
+    container.appendChild(appRoot);
+
+    // Placeholder de chargement
+    const loader = document.createElement('div');
+    loader.id = 'tto-loading';
+    loader.textContent = 'Chargement...';
+    appRoot.appendChild(loader);
+
+    document.body.appendChild(container);
+
+    // Lancer le slide-in après un délai, pendant que React charge
+    setTimeout(() => {
+      requestAnimationFrame(() => {
+        container.style.transform = 'translateX(0%)';
+      });
+    }, ANIMATION_DELAY);
+
+    const script = document.createElement('script');
+    script.src = chrome.runtime.getURL('main.js') + '?t=' + Date.now();
+    script.type = 'module';
+    script.onload = () => {
+      const l = document.getElementById('tto-loading');
+      if (l) {
+        l.remove();
+      }
+      if (window.initTTO) {
+        window.initTTO();
+      }
+      setTimeout(() => {
+        document.dispatchEvent(new CustomEvent('TTO_PANEL_OPENED'));
+      }, 100);
+    };
+    document.head.appendChild(script);
+  }
+
+  // Ferme le panneau
+  function closePanel(container) {
+    if (window.__ttoRoot) { try { window.__ttoRoot.unmount(); } catch {} }
+    container.style.transition = 'transform 0.20s ease-in';
+    container.style.transform = 'translateX(100%)';
+    container.addEventListener('transitionend', () => {
+      container.remove();
+      document.dispatchEvent(new CustomEvent('TTO_PANEL_CLOSED'));
+    }, { once: true });
+  }
+
+  // Toggle selon existence
+  function togglePanel() {
+    const existing = document.getElementById(containerId);
+    if (existing) closePanel(existing);
+    else openPanel();
+  }
+
+  // Exécution du toggle
+  togglePanel();
+}
